refactor: drop canvas_id claim and consolidate OIDC verify code
Remove canvas_id from Semaphore OIDC JWT claims and merge remote
verification helpers into verify.go.
